I found the bug. It was not about numerical errors but extracting the correct values for the input gradients (dx). It occurred that I did not shifted the values correctly according to padding once the returned values from inverse FFT had to be cropped.
I went thorough the training of ResNet-18 and applied padding and strides. For one of the cases, the added padding made my input gradients incorrect, though the filter gradients were correct.

Now, for a small dataset, we have almost an exact match of the pytorch's conv2D and our FFT-based conv2D.

I will train the network on the full CIFAR-10 dataset.